Add confirmation dialog on delete all, Change button styling

+++ modules/testing/src/views/full/index.jsx
@@ -4,6 +4,7 @@ import { Icon, Intent } from '@blueprintjs/core'
 import { IconNames } from '@blueprintjs/icons'
 import style from './style.scss'
 import ScenarioRecorder from './ScenarioRecorder'
+import { confirmDialog } from 'botpress/shared'
 import NoScenarios from './NoScenarios'
 import Scenario from './Scenario'
 
@@ -48,8 +49,13 @@ export default class Testing extends React.Component {
   }
 
   deleteAllScenarios = async () => {
-    await this.props.bp.axios.post('/mod/testing/deleteAllScenarios')
-    await this.loadScenarios()
+    const shouldDelete = await confirmDialog('Are you sure you want to delete all of the scenarios?', {
+      acceptLabel: 'Delete all'
+    })
+    if (shouldDelete) {
+      await this.props.bp.axios.post('/mod/testing/deleteAllScenarios')
+      await this.loadScenarios()
+    }
   }
 
   loadPreviews = async () => {
@@ -159,29 +165,26 @@ export default class Testing extends React.Component {
             <Col md={10} mdOffset={1}>
               <Row>
                 {/* TODO extract this in header component ? */}
-                <Col md={8}>
-                  <h2>Scenarios</h2>
-                  {this.renderSummary()}
-                </Col>
-                {!this.state.isRecording && (
-                  <Col md={4}>
-                    <div className="pull-right">
-                      <Button bsSize="small" onClick={this.runAllScenarios} disabled={this.state.isRunning}>
-                        <Icon icon={IconNames.PLAY} intent={Intent.SUCCESS} /> Run All
-                      </Button>
-                      &nbsp;
-                      <Button bsSize="small" onClick={this.startRecording}>
-                        <Icon icon={IconNames.RECORD} /> Record new
-                      </Button>
-                      &nbsp;
-                      <Button bsSize="small" variant="danger" onClick={this.deleteAllScenarios}>
-                        <Icon icon={IconNames.TRASH} /> Delete all
-                      </Button>
-                    </div>
-                  </Col>
-                )}
+                <h2>Scenarios</h2>
+                {this.renderSummary()}
               </Row>
-
+              {!this.state.isRecording && (
+                <Row className={style['actions-container']}>
+                  <Button bsSize="small" variant="danger" className={'btn-danger'} onClick={this.deleteAllScenarios}>
+                    <Icon icon={IconNames.TRASH} /> Delete all
+                  </Button>
+                  <Button bsSize="small" onClick={this.runAllScenarios} disabled={this.state.isRunning}>
+                    <Icon icon={IconNames.PLAY} intent={Intent.SUCCESS} /> Run All
+                  </Button>
+                  <Button bsSize="small" onClick={this.startRecording}>
+                    <Icon icon={IconNames.RECORD} /> Record new
+                  </Button>
+                </Row>
+              )}
+            </Col>
+          </Row>
+          <Row>
+            <Col md={10} mdOffset={1}>
               {this.state.isRecording && (
                 <ScenarioRecorder
                   bp={this.props.bp}
